backend/requirements.txt
asgiref==3.11.1
Django==5.1.7
djangorestframework==3.17.0
requests==2.32.3
sqlparse==0.5.5
